splitting tsx and schema files
